Revert "CR-17813 (#800)"
This reverts commit acbce0131ef655a62d8176dccfb5b4593cfe780f.
