+++ docs/content/pipelines/Run Pipeline.md
@@ -15,44 +15,22 @@ The pipeline will be triggered multiple times according to the array length.
 
 #### Variable yaml file with 2 sets of variables
 ```yaml
-- VARIABLE_A: value_a_for_the_first_build
-  VARIABLE_B: value_b_for_the_first_build
-- VARIABLE_A: value_a_for_the_first_build
-  VARIABLE_B: value_b_for_the_first_build
+- key: value
+  key2: key1
+- key: value
+  key2: key2
 ```
 
 #### Variable json file with 2 sets of variables
 ```json
 [
   {
-    "VARIABLE_A": "value_a_for_the_first_build",
-    "VARIABLE_B": "value_b_for_the_first_build"
+    "key": "value",
+    "key2": "key1"
   },
   {
-    "VARIABLE_A": "value_a_for_the_first_build",
-    "VARIABLE_B": "value_b_for_the_first_build"
-  }
-]
-```
-### Use encrypted variables in Codefresh build runs; supported from CLI version: 0.82.8
-#### Variable yaml file with single variable set with encrypted variables
-```yaml
-- key:
-    val: value
-    encrypted: true
-  key2: key1
-
-```
-
-#### Variable json file single variable set with encrypted variables
-```json
-[
-  {
-    "key": {
-      "val": "value",
-      "encrypted": true
-    },
-    "key2": "key1"
+    "key": "value",
+    "key2": "key2"
   }
 ]
 ```
