+++ lib/interface/cli/commands/pipeline/pipeline.sdk.spec.js
@@ -1,6 +1,3 @@
-const yaml = require('js-yaml');
-const request = require('requestretry');
-const fs = require('fs');
 const DEFAULTS = require('../../defaults');
 const getCmd = require('./get.cmd').toCommand();
 const deleteCmd = require('./delete.cmd').toCommand();
@@ -13,15 +10,17 @@ jest.mock('../../helpers/validation'); // eslint-disable-line
 jest.mock('../../../../../check-version');
 jest.mock('../../completion/helpers', () => { // eslint-disable-line
     return {
-        authContextWrapper: (func) => func,
+        authContextWrapper: func => func,
     };
 });
 
 jest.mock('../../../../logic/entities/Pipeline', () => { // eslint-disable-line
     return {
-        fromResponse: (res) => res,
+        fromResponse: res => res,
     };
 });
+
+const request = require('requestretry');
 
 const DEFAULT_RESPONSE = request.__defaultResponse();
 
@@ -58,11 +57,11 @@ describe('pipeline', () => {
             });
 
             it('should return default limit', async () => {
-                expect(_getLimit(undefined, false)).toEqual(DEFAULTS.GET_LIMIT_RESULTS);
+                expect(_getLimit(undefined,false)).toEqual(DEFAULTS.GET_LIMIT_RESULTS);
             });
 
             it('should return `unlimited` value', async () => {
-                expect(_getLimit(undefined, true)).toEqual(DEFAULTS.GET_ALL_PIPELINES_LIMIT);
+                expect(_getLimit(undefined,true)).toEqual(DEFAULTS.GET_ALL_PIPELINES_LIMIT);
             });
         });
 
@@ -82,75 +81,6 @@ describe('pipeline', () => {
                         const pip = new BasePipeline(argv);
                         await pip.preRunAll();
                         await verifyResponsesReturned([DEFAULT_RESPONSE, DEFAULT_RESPONSE]); // eslint-disable-line
-                    });
-                });
-
-                describe('run', () => {
-                    it('should handle running pipeline with encrypted variables', async () => {
-                        const argv = { name: 'some name',
-                            detach: true,
-                            annotation: [],
-                            variable:
-                                [{
-                                    key: 'secret',
-                                    value: 'secret',
-                                },
-                                {
-                                    key: 'VAR1',
-                                    value: 'VAL1',
-                                },
-                                ],
-                            encrypted: ['secret'],
-                        };
-                        const pip = new CfPipeline(argv);
-                        await pip.run();
-                        expect(pip.executionRequests[0].options.variables).toEqual([
-                            {
-                                key: 'secret',
-                                value: 'secret',
-                                encrypted: true,
-                            },
-                            {
-                                key: 'VAR1',
-                                value: 'VAL1',
-                            },
-                        ]);
-                        await verifyResponsesReturned([DEFAULT_RESPONSE]); // eslint-disable-line
-                    });
-
-                    it('should handle running pipeline with encrypted variables passing inside json file', async () => {
-                        const rawFile = fs.readFileSync('lib/interface/cli/commands/pipeline/var.json', 'utf8');
-
-                        const argv = { name: 'some name',
-                            detach: true,
-                            annotation: [],
-                            'var-file': JSON.parse(rawFile),
-                        };
-                        const pip = new CfPipeline(argv);
-                        await pip.run();
-                        expect(pip.executionRequests[0].options.variables).toEqual(
-                            [{ key: 'help6', value: '85858' },
-                                { key: 'should_be_encrepted', value: '0000' },
-                                { encrypted: true, key: 'help7', value: 'test' }],
-                        );
-                        await verifyResponsesReturned([DEFAULT_RESPONSE]); // eslint-disable-line
-                    });
-
-                    it('should handle running pipeline with encrypted variables passing inside yaml file', async () => {
-                        const rawFile = fs.readFileSync('lib/interface/cli/commands/pipeline/var.yml', 'utf8');
-
-                        const argv = { name: 'some name',
-                            detach: true,
-                            annotation: [],
-                            'var-file': yaml.safeLoad(rawFile),
-                        };
-                        const pip = new CfPipeline(argv);
-                        await pip.run();
-                        expect(pip.executionRequests[0].options.variables).toEqual(
-                            [{ key: 'VAR1', value: 'VAL1' },
-                                { encrypted: true, key: 'VAR2', value: 'VAL2' }],
-                        );
-                        await verifyResponsesReturned([DEFAULT_RESPONSE]); // eslint-disable-line
                     });
                 });
