+++ lib/interface/cli/commands/pipeline/run.base.js
@@ -1,11 +1,9 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
-const CFError = require('cf-errors');
-const { prepareKeyValueFromCLIEnvOption,
-    markEncryptedFlagOnRequestedVariables,
-    prepareKeyValueObjectsFromEnvFileOption } = require('../../helpers/general');
+const { prepareKeyValueFromCLIEnvOption } = require('../../helpers/general');
 const { validatePipelineYaml } = require('../../helpers/validation');
 const { printResult } = require('../root/validate.cmd');
+const CFError = require('cf-errors');
 const { sdk } = require('../../../../logic');
 
 class RunBaseCommand {
@@ -61,18 +59,18 @@ class RunBaseCommand {
         if (variablesFromFile) {
             _.forEach(variablesFromFile, (variables) => {
                 const request = _.cloneDeep(executionRequestTemplate);
-                request.options.variables = prepareKeyValueObjectsFromEnvFileOption(variables);
+                request.options.variables = variables;
                 this.executionRequests.push(request);
             });
         } else {
-            const variables = markEncryptedFlagOnRequestedVariables(this.argv.variable, this.argv.encrypted);
+            const variables = prepareKeyValueFromCLIEnvOption(this.argv.variable);
             const request = _.cloneDeep(executionRequestTemplate);
             request.options.variables = variables;
             request.options.contexts = contexts;
             this.executionRequests.push(request);
         }
 
-        const results = await Promise.all(this.executionRequests.map((request) => this.runImpl(request)));
+        const results = await Promise.all(this.executionRequests.map(request => this.runImpl(request)));
         const findMaxReducer = (accumulator, currentValue) => (currentValue > accumulator ? currentValue : accumulator);
         const exitCode = results.reduce(findMaxReducer);
         await this.postRunRequest();
