+++ lib/interface/cli/commands/pipeline/run.cmd.js
@@ -1,5 +1,6 @@
+const debug = require('debug')('codefresh:cli:run:pipeline');
 const Command = require('../../Command');
-const { crudFilenameOption, prepareKeyValueObjectsFromCLIEnvOption } = require('../../helpers/general');
+const { crudFilenameOption } = require('../../helpers/general');
 const RunLocalCommand = require('./run.local');
 const RunExternalCommand = require('./run.cf');
 
@@ -84,13 +85,6 @@ const run = new Command({
                 describe: 'Set build variables',
                 default: [],
                 alias: 'v',
-                coerce: prepareKeyValueObjectsFromCLIEnvOption,
-            })
-            .option('encrypted', {
-                array: true,
-                alias: 'e',
-                describe: 'Variable names to encrypt',
-                default: [],
             })
             .option('detach', {
                 alias: 'd',
@@ -132,7 +126,6 @@ const run = new Command({
             .example('codefresh run PIPELINE_ID | PIPELINE_NAME -b=master', 'Defining the source control context using a branch')
             .example('codefresh run PIPELINE_ID | PIPELINE_NAME -s=52b992e783d2f84dd0123c70ac8623b4f0f938d1', 'Defining the source control context using a commit')
             .example('codefresh run PIPELINE_ID | PIPELINE_NAME -b=master -v key1=value1 -v key2=value2', 'Setting variables through the command')
-            .example('codefresh run PIPELINE_ID | PIPELINE_NAME -b=master -v key1=value1 -v key2=value2 -e key1', 'Setting variables through the command with encrypted option')
             .example('codefresh run PIPELINE_ID | PIPELINE_NAME -b=master --var-file ./var_file.yml', 'Settings variables through a yml file')
             .example('codefresh run PIPELINE_ID | PIPELINE_NAME -b=master --context context', 'Inject contexts to the pipeline execution')
             .example('codefresh run PIPELINE_ID | PIPELINE_NAME --skip step1 step2 step3', 'Skip specific steps');
