--- lib/interface/cli/commands/pipeline/var.json
@@ -1,12 +1,0 @@
-{
-  "build1": {
-    "help6": "85858",
-    "should_be_encrepted": "0000",
-    "help7": {
-      "value": "test",
-      "encrypted": true
-    }
-  }
-}
-
-
