--- lib/interface/cli/commands/pipeline/var.yml
@@ -1,5 +1,0 @@
-build1:
-  VAR1: 'VAL1'
-  VAR2:
-    value: VAL2
-    encrypted: true
