+++ lib/interface/cli/commands/project/apply.cmd.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const { sdk } = require('../../../../logic');
 
 const applyRoot = require('../root/apply.cmd');
-const { prepareKeyValueObjectsFromCLIEnvOption, ignoreHttpError, markEncryptedFlagOnRequestedVariables } = require('../../helpers/general');
+const { prepareKeyValueObjectsFromCLIEnvOption, ignoreHttpError } = require('../../helpers/general');
 
 const command = new Command({
     command: 'project [id|name]',
@@ -61,7 +61,14 @@ const command = new Command({
             encrypted,
         } = argv;
 
-        const requestedProjectVariables = markEncryptedFlagOnRequestedVariables(variables, encrypted);
+        const variableMap = _.reduce(variables, (acc, v) => _.assign(acc, { [v.key]: v }), {});
+        _.forEach(encrypted, (varName) => {
+            const variable = variableMap[varName];
+            if (!variable) {
+                throw new CFError(`Variable is not provided: "${varName}"`);
+            }
+            variable.encrypted = true;
+        });
 
         let project = await sdk.projects.get({ id }).catch(ignoreHttpError);
         project = project || await sdk.projects.getByName({ name }).catch(ignoreHttpError);
@@ -74,7 +81,7 @@ const command = new Command({
         const updatePayload = _.pickBy({
             projectName,
             tags: tags || existingTags,
-            variables: requestedProjectVariables || existingVariables,
+            variables: variables || existingVariables,
         }, _.identity);
 
         await sdk.projects.patch({ id: project.id }, updatePayload);
