+++ lib/interface/cli/commands/project/create.cmd.js
@@ -1,8 +1,10 @@
 const Command = require('../../Command');
+const CFError = require('cf-errors');
+const _ = require('lodash');
 const { sdk } = require('../../../../logic');
 const createRoot = require('../root/create.cmd');
 const { checkOrProjectExists } = require('../../helpers/validation');
-const { prepareKeyValueObjectsFromCLIEnvOption, markEncryptedFlagOnRequestedVariables } = require('../../helpers/general');
+const { prepareKeyValueObjectsFromCLIEnvOption } = require('../../helpers/general');
 
 const command = new Command({
     command: 'project <name>',
@@ -50,10 +52,17 @@ const command = new Command({
             encrypted,
         } = argv;
 
-        const requestedProjectVariables = markEncryptedFlagOnRequestedVariables(variables, encrypted);
+        const variableMap = _.reduce(variables, (acc, v) => _.assign(acc, { [v.key]: v }), {});
+        _.forEach(encrypted, (varName) => {
+            const variable = variableMap[varName];
+            if (!variable) {
+                throw new CFError(`Variable is not provided: "${varName}"`);
+            }
+            variable.encrypted = true;
+        });
 
         await checkOrProjectExists(projectName);
-        await sdk.projects.create({ projectName, tags, variables: requestedProjectVariables });
+        await sdk.projects.create({ projectName, tags, variables });
         console.log(`Project: "${projectName}" created`);
     },
 });
