+++ lib/interface/cli/helpers/general.js
@@ -2,9 +2,9 @@ const Promise = require('bluebird');
 const _ = require('lodash');
 const fs = require('fs');
 const yaml = require('js-yaml');
+const defaults = require('../defaults');
 const CFError = require('cf-errors');
 const path = require('path');
-const defaults = require('../defaults');
 const Output = require('../../../output/Output');
 const { sdk } = require('../../../logic');
 
@@ -106,39 +106,6 @@ const prepareKeyValueObjectsFromCLIEnvOption = (environmentVariables) => {
     return variables;
 };
 
-/**
- * will return an array of objects { key, value, encrypted } from parsing an array of objects: key=value ,key: {value: "value", encrypted: true }
- * @param environmentVariables
- * @returns Array of { key, value, encrypted }
- */
-const prepareKeyValueObjectsFromEnvFileOption = (environmentVariables) => _.map(environmentVariables, (value, key) => {
-    if (_.isObject(value)) {
-        const { value: val, encrypted } = value;
-        return { key, value: val, encrypted };
-    }
-    return { key, value };
-});
-
-/**
- * The function takes in a list of requested variables and an array of encrypted variable names,
- * and sets the encrypted flag to true on each corresponding variable object in the list.
- * @param environmentVariables
- * @param encrypted
- * @returns Array of { key, value, encrypted }
- */
-const markEncryptedFlagOnRequestedVariables = (requestedVariables, encrypted) => {
-    const variables = _.cloneDeep(requestedVariables);
-    const variableMap = _.reduce(variables, (acc, v) => _.assign(acc, { [v.key]: v }), {});
-    _.forEach(encrypted, (varName) => {
-        const variable = variableMap[varName];
-        if (!variable) {
-            throw new CFError(`Variable is not provided: "${varName}"`);
-        }
-        variable.encrypted = true;
-    });
-    return variables;
-};
-
 const crudFilenameOption = (yargs, options = {}) => {
     const filenameOption = {
         alias: options.alias || 'f',
@@ -155,19 +122,13 @@ const crudFilenameOption = (yargs, options = {}) => {
         .coerce(options.name || 'filename', (arg) => {
             try {
                 const rawFile = fs.readFileSync(path.resolve(process.cwd(), arg), 'utf8');
-                let content;
                 if (arg.endsWith('.json')) {
-                    content = options.raw ? rawFile : JSON.parse(rawFile);
-                } else if (arg.endsWith('.yml') || arg.endsWith('yaml')) {
-                    content = options.raw ? rawFile : yaml.safeLoad(rawFile);
-                } else {
-                    throw new CFError('File extension is not recognized');
+                    return options.raw ? rawFile : JSON.parse(rawFile);
                 }
-                if (_.isObject(content)) {
-                    return content;
+                if (arg.endsWith('.yml') || arg.endsWith('yaml')) {
+                    return options.raw ? rawFile : yaml.safeLoad(rawFile);
                 }
-                throw new CFError('Not a valid file.\n'
-                    + 'For more information how to pass a valid file, go to https://codefresh-io.github.io/cli/pipelines/run-pipeline/');
+                throw new CFError('File extension is not recognized');
             } catch (err) {
                 const error = new CFError({
                     message: 'Failed to read file',
@@ -200,15 +161,15 @@ const selectColumnsOption = (yargs, options = {}) => {
 };
 
 function pathExists(p) {
-    return new Promise((resolve) => fs.access(p, resolve))
-        .then((err) => !err);
+    return new Promise(resolve => fs.access(p, resolve))
+        .then(err => !err);
 }
 
 const readFile = Promise.promisify(fs.readFile);
 
 function watchFile(filename, listener) {
     fs.watchFile(filename, { interval: 500 }, listener);
-    const unwatcher = (f) => () => fs.unwatchFile(f);
+    const unwatcher = f => () => fs.unwatchFile(f);
     ['exit', 'SIGINT', 'SIGUSR1', 'SIGUSR2', 'uncaughtException', 'SIGTERM'].forEach((eventType) => {
         process.on(eventType, unwatcher(filename));
     });
@@ -216,7 +177,7 @@ function watchFile(filename, listener) {
 
 function ignoreHttpError(e) {
     if (!e.statusCode) {
-        throw e;
+        throw e
     }
     return undefined;
 }
@@ -271,6 +232,4 @@ module.exports = {
     detectProxy,
     keyValueArrayToObject,
     addProxyVariables,
-    markEncryptedFlagOnRequestedVariables,
-    prepareKeyValueObjectsFromEnvFileOption,
 };
