+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "codefresh",
-  "version": "0.82.8",
+  "version": "0.82.7",
   "description": "Codefresh command line utility",
   "main": "index.js",
   "preferGlobal": true,
